Fixed stylesheet error
The argument passed to the url function should not be quoted, at least in the case of an image file
from the MadCap Flare software installation.
An incorrect value causes a silent warning (error code 3, without actual warning printed to the log file),
and thus an error while building the documentation target during releases.